Rename ITEM_TYPE to ItemType, turn it into an enum class
Free-standing rather than nested in XItem - same wall as ItemKind:
item/itemdb.h's _MAIN_ITEM_STRUCT::it needs it before class XItem is
declared. Unlike ItemKind/AttackEffectType, this one is a plain
sequential enum (not a bitmask - every value is a distinct item
category, never combined via | or tested via &), confirmed by grepping
for bitwise ops on `it`/ITEM_TYPE values before starting: found none,
so no operator overloads were needed at all, just the mechanical
qualification.

+++ game/location.cpp
@@ -952,7 +952,7 @@ XGUID XLocation::GetObjectGUID(void* obj)
 std::tuple<int, int, int, int, int, std::string> XLocation::GetItemParam(void* item)
 {
     XItem * p = (XItem*)item;
-    return {static_cast<int>(p->kind), static_cast<int>(p->aet), p->wt, p->it, p->quantity, p->name};
+    return {static_cast<int>(p->kind), static_cast<int>(p->aet), p->wt, static_cast<int>(p->it), p->quantity, p->name};
 }
 
 void XLocation::SetItemBrand(void* item, int br)
